mrc-kv-cache-dtype(extract-kv-cache-dtype-move,mechanical_provable): Move configure_kv_cache_dtype + TORCH_DTYPE_TO_KV_CACHE_STR to mem_cache.kv_cache_dtype (cut+paste)
