+++ python/sglang/srt/model_executor/model_runner.py
@@ -31,7 +31,6 @@ from typing import Any, Callable, List, Optional, Tuple, Union
 import torch
 import torch.distributed as dist
 
-from sglang.kernels.ops.quantization.fp8_kernel import fp8_dtype
 from sglang.srt.configs import (
     BailingHybridConfig,
     FalconH1Config,
@@ -140,6 +139,7 @@ from sglang.srt.lora.lora_manager import LoRAManager
 from sglang.srt.lora.lora_registry import LoRARef
 from sglang.srt.managers.schedule_batch import sanity_check_mm_pad_shift_value
 from sglang.srt.mem_cache.allocator import BaseTokenToKVPoolAllocator
+from sglang.srt.mem_cache.kv_cache_dtype import configure_kv_cache_dtype
 from sglang.srt.mem_cache.memory_pool import HybridReqToTokenPool, ReqToTokenPool
 from sglang.srt.model_executor.cpu_graph_runner import CPUGraphRunner
 from sglang.srt.model_executor.cuda_graph_config import (
@@ -258,13 +258,6 @@ if _is_npu:
     init_npu_backend()
 elif current_platform.is_out_of_tree():
     current_platform.init_backend()
-
-TORCH_DTYPE_TO_KV_CACHE_STR = {
-    torch.float8_e4m3fn: "fp8_e4m3",
-    torch.float8_e4m3fnuz: "fp8_e4m3",
-    torch.float8_e5m2: "fp8_e5m2",
-    torch.bfloat16: "bf16",
-}
 
 
 # Detect stragger ranks in model loading
@@ -775,12 +768,10 @@ class ModelRunner(ModelRunnerKVCacheMixin):
 
             enable_batch_invariant_mode()
 
-        resolved_kv_cache_dtype, self.kv_cache_dtype = (
-            ModelRunner.configure_kv_cache_dtype(
-                server_args=self.server_args,
-                model=self.model,
-                model_dtype=self.dtype,
-            )
+        resolved_kv_cache_dtype, self.kv_cache_dtype = configure_kv_cache_dtype(
+            server_args=self.server_args,
+            model=self.model,
+            model_dtype=self.dtype,
         )
         if resolved_kv_cache_dtype is not None:
             self._record_kv_cache_dtype(resolved_kv_cache_dtype)
@@ -2386,53 +2377,6 @@ class ModelRunner(ModelRunnerKVCacheMixin):
             self.server_args.override(
                 "ModelRunner.configure_kv_cache_dtype", kv_cache_dtype=resolved
             )
-
-    @staticmethod
-    def configure_kv_cache_dtype(
-        *,
-        server_args: ServerArgs,
-        model: nn.Module,
-        model_dtype: torch.dtype,
-    ) -> tuple[Optional[str], torch.dtype]:
-        resolved_kv_cache_dtype: Optional[str] = None
-        server_args_kv_cache_dtype = server_args.kv_cache_dtype
-        if server_args_kv_cache_dtype == "auto":
-            quant_config = getattr(model, "quant_config", None)
-            kv_cache_quant_algo = getattr(quant_config, "kv_cache_quant_algo", None)
-            if (
-                isinstance(kv_cache_quant_algo, str)
-                and kv_cache_quant_algo.upper() == "FP8"
-            ):
-                kv_cache_dtype = fp8_dtype if _is_hip else torch.float8_e4m3fn
-                resolved_kv_cache_dtype = TORCH_DTYPE_TO_KV_CACHE_STR[kv_cache_dtype]
-            else:
-                kv_cache_dtype = model_dtype
-        elif server_args_kv_cache_dtype == "fp8_e5m2":
-            if _is_hip:  # Using natively supported format
-                kv_cache_dtype = fp8_dtype
-            else:
-                kv_cache_dtype = torch.float8_e5m2
-        elif server_args_kv_cache_dtype == "fp8_e4m3":
-            if _is_hip:  # Using natively supported format
-                kv_cache_dtype = fp8_dtype
-            else:
-                kv_cache_dtype = torch.float8_e4m3fn
-        elif server_args_kv_cache_dtype in ("bf16", "bfloat16"):
-            kv_cache_dtype = torch.bfloat16
-        elif server_args_kv_cache_dtype == "fp4_e2m1":
-            if hasattr(torch, "float4_e2m1fn_x2"):
-                kv_cache_dtype = torch.float4_e2m1fn_x2
-                logger.warning(f"FP4 (E2M1) KV Cache might lead to a accuracy drop!")
-            else:
-                logger.warning(
-                    f"--kv-cache-dtype falls back to 'auto' because this torch version does not support torch.float4_e2m1fn_x2"
-                )
-                kv_cache_dtype = model_dtype
-        else:
-            raise ValueError(
-                f"Unsupported kv_cache_dtype: {server_args_kv_cache_dtype}."
-            )
-        return resolved_kv_cache_dtype, kv_cache_dtype
 
     def init_attention_backend(self):
         """Init attention kernel backend."""
